Section: A queueing model with 2 consecutive buffer centres (#6)
* Add bibliography entry

* Write details of the Markov-chain model

- Define the model
- Define state space
- Define transition matrix and state-mapping function

* Add general Markov-chain picture

* Update known.py

* Add .pdf file to GitHub repo

* Create subsections/ subdirectory

* Minor edits on figures

Reference figure of previous section and change the caption of
the general Markov chain figure

* Add paragraph for transition rates

* Apply minor edits

* Write main.tex for performance measures subsection

* Write main.tex for waiting time subsubsection

* Update known.py

* Rename directory

* Fix typo

* Apply suggestions from code review

Co-authored-by: Vince Knight <vincent.knight@gmail.com>

* Apply more siggestions from review

- Markov chains intro
- State space definition
- Replace q_{i,j} with Q_{ij}
- Minor changes

* Update known.py

* Minor edits

* Add blocking time algorithm

* Add python code that generates tikz picture

In order for the code to work, ambulance_game needs
to be imported first.

* Place all tikz pictures in imgs\ directory

* Remove some un-needed comments

* Update known.py

* Add another markov chain example plot

* Apply minor changes to some .tex files

* Add proportion within target subsubsection

Add subsubsection that explains how the proportion of individuals
is calculated. There are several subparagraphs to that section that
explain additional components that may or may not be left in at the
end.

* Update CI

* Comment out section

* Change bibliography and comment mathtools package

* Use plain bibliography style

* Add known words

* Use mathtools package

* Update CI and fix typo

Remove unwanted tex libraries

* Include section that was commented out

* Minor change on input

* Comment most of proportion section out

* Undo last commit

* Commit .pdf as well

* Apply suggestions from code review

Co-authored-by: Vince Knight <vincent.knight@gmail.com>

* Change the way equations are cited

Include brackets around cited equations

* Replace the word "class" with "type"

Class 1 and class 2 individulas are now referred to as
type 1 and type 2 individuals

* Apply suggestions to waiting time section from code review

Co-authored-by: Vince Knight <vincent.knight@gmail.com>

* Restructure waiting time formula section

The section is now structured in the
following way:

- Formula 1
- Formula 2
- Formula 3
- Defining necessary parameters

* Minor change in notation on waiting time formula

* Restructure blocking time formula

Remove most of text that focuses on the "how" part of the formula.
What's left now is just the formula and everything that's necessary to
use it.

* Add reference for sojourn time

* Add an appendix

Currently the appendix has some information of how to use the blocking
time formula. Soon to be added are information on the waiting time
formula (possibly the recursive method) and information on the
proportion within target formula.

* Restructure proportion within time target formula

Remove most parts of this section and move them to appendix

* Add python code to generate violinplots

Added 3 python files that will generate violinplots for
the waiting time, blocking time and proportion within target
using the simulation model. These are then compared in the
same plot with the analytical outcomes using the Markov
chains method.

* Add comparisons between Markov and DES

Include violinplots of waiting time, blocking time and proportion
within target time. The violinplots show the differences between
Markov values and simulated values.

* Add new plots to compare Markov and DES

The comparisons are now made over different values of
lambda_2, for all performance measures for all possible
types of individuals.

* Add a todo to add description for DES

* Fix latex compile issue with pdfs (#8)

* Remove scalebox before \includegraphics

* Add updated .pdf

* Chnage the way pdf compiles in CI

tasks.py are updated so that latexmk runs with the xelatex parameter

* Update CI to install additional libraries

Specifically: texlive-xetex, texlive-science

* Restructure imgs/ directory

* Add new violinplots

Include violin plots of performance measures over
different values of lambda_2 for type 1, type 2 and
overall individuals

* Apply changes from code review

* Restructure queueing model section

Remove performance measures subsection and change "waiting time",
"blocking time" and "proportion of inds" from subsubsections into
subsections. Also add example subsubsections

* Write appendix chapter for DES model

* Edit minor things

* Add recursive waiting time formula to appendix

* Minor change

* Apply minor changes from review

Co-authored-by: Vince Knight <vincent.knight@gmail.com>